+++ src/coreclr/jit/importercalls.cpp
@@ -3342,7 +3342,7 @@ GenTree* Compiler::impIntrinsic(CORINFO_CLASS_HANDLE    clsHnd,
         return node;
     }
 
-    if (ni == NI_System_Runtime_CompilerServices_RuntimeHelpers_AsyncSuspend)
+    if (ni == NI_System_Runtime_CompilerServices_AsyncHelpers_AsyncSuspend)
     {
         GenTree* node = gtNewOperNode(GT_RETURN_SUSPEND, TYP_VOID, impPopStack().val);
         node->SetHasOrderingSideEffect();
@@ -11022,14 +11022,17 @@ NamedIntrinsic Compiler::lookupNamedIntrinsic(CORINFO_METHOD_HANDLE method)
                             {
                                 result = NI_System_Runtime_CompilerServices_RuntimeHelpers_Await;
                             }
-                            else if (strcmp(methodName, "AsyncSuspend") == 0)
-                            {
-                                result = NI_System_Runtime_CompilerServices_RuntimeHelpers_AsyncSuspend;
-                            }
                             else if (strcmp(methodName, "get_RuntimeAsyncViaJitGeneratedStateMachines") == 0)
                             {
                                 result =
                                     NI_System_Runtime_CompilerServices_RuntimeHelpers_get_RuntimeAsyncViaJitGeneratedStateMachines;
+                            }
+                        }
+                        if (strcmp(className, "AsyncHelpers") == 0)
+                        {
+                            if (strcmp(methodName, "AsyncSuspend") == 0)
+                            {
+                                result = NI_System_Runtime_CompilerServices_AsyncHelpers_AsyncSuspend;
                             }
                         }
                         else if (strcmp(className, "StaticsHelpers") == 0)
