+++ src/coreclr/jit/namedintrinsiclist.h
@@ -120,8 +120,9 @@ enum NamedIntrinsic : unsigned short
     NI_System_Runtime_CompilerServices_RuntimeHelpers_IsReferenceOrContainsReferences,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_GetMethodTable,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_Await,
-    NI_System_Runtime_CompilerServices_RuntimeHelpers_AsyncSuspend,
     NI_System_Runtime_CompilerServices_RuntimeHelpers_get_RuntimeAsyncViaJitGeneratedStateMachines,
+
+    NI_System_Runtime_CompilerServices_AsyncHelpers_AsyncSuspend,
 
     NI_System_Runtime_CompilerServices_StaticsHelpers_VolatileReadAsByref,
